Add `PyBlockExecutor` (#776)
* For simplicity, all it does at the moment is hold general_config and
  max_recursion_depth, where the former's definition was moved into
  `PyBlockExecutor`'s module, since it is its primary owner.
* It is also in charge of initializing the PyTransactionExecutor for
  Python.
* Make `py_block_context` into a method of `PyGeneralConfig` --- note
  that PyBlockInfo can be used there instead of PyAny, will do in
  separate PR to keep the diff small.
* in subsequent PRs it will also be in charge of Storage, and later on
  hide both storage and transaction executor from Python.
