+++ crates/native_blockifier/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod errors;
 pub mod papyrus_state;
+pub mod py_block_executor;
 pub mod py_state_diff;
 #[cfg(any(feature = "testing", test))]
 pub mod py_test_utils;
@@ -10,6 +11,7 @@ pub mod py_utils;
 pub mod storage;
 
 use errors::add_py_exceptions;
+use py_block_executor::PyBlockExecutor;
 use py_transaction_execution_info::{
     PyCallInfo, PyOrderedEvent, PyOrderedL2ToL1Message, PyTransactionExecutionInfo,
     PyVmExecutionResources,
@@ -27,6 +29,7 @@ fn native_blockifier(py: Python<'_>, py_module: &PyModule) -> PyResult<()> {
     // Usage: just create a Python logger as usual, and it'll capture Rust prints.
     pyo3_log::init();
 
+    py_module.add_class::<PyBlockExecutor>()?;
     py_module.add_class::<PyCallInfo>()?;
     py_module.add_class::<PyOrderedEvent>()?;
     py_module.add_class::<PyOrderedL2ToL1Message>()?;
